third_party/blink/web_tests/fast/dom/HTMLElement/class-list-feature-detection-throwing.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
    test(function() {
        var div = document.createElement("div");
        // Test that "supports()" is throwing TypeError
        assert_throws_js(
            TypeError, function() { div.classList.supports("bogus"); },
            'supports method should throw if called on an DOMTokenList with no supported tokens, such as classList');
    }, "Make sure that classList throws when supports() is called");
</script>
